backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Igorik <egorsborsikov250575@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
supabase = "^2.15.1"
pydantic = "^2.11.4"
httpx = "^0.28.1"
uvicorn = "^0.34.2"
redis = "^6.0.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
